Finished writing RESTful interactions

+++ src/main/java/edu/greenriver/sdev/saasproject/controllers/BoardGameApi.java
@@ -7,6 +7,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 public class BoardGameApi {
@@ -18,31 +19,38 @@ public class BoardGameApi {
 
     @GetMapping("boardgames")
     public ResponseEntity<List<BoardGame>> allBoardGames(){
-        return new ResponseEntity<>(bgservice.getAllBoardGames(), HttpStatus.OK);
+        return new ResponseEntity<>(bgservice.getAllBoardGames(), HttpStatus.OK);//200
     }
 
     @GetMapping("boardgames/random")
     public ResponseEntity<BoardGame> random(){
-        return new ResponseEntity<>(bgservice.random(), HttpStatus.OK);
+        return new ResponseEntity<>(bgservice.random(), HttpStatus.OK);//200
     }
 
     @GetMapping("boardgames/{bgId}")
     public ResponseEntity<BoardGame> getBoardGameById(@PathVariable int bgId){
-        return new ResponseEntity<>(bgservice.getBoardGameById(bgId),HttpStatus.OK);
+        return new ResponseEntity<>(bgservice.getBoardGameById(bgId),HttpStatus.OK);//200
     }
 
     @PostMapping("boardgames")
     public ResponseEntity<BoardGame> addBoardGame(@RequestBody BoardGame boardGame){
-        return new ResponseEntity<>(bgservice.addBoardGame(boardGame), HttpStatus.CREATED);
+        return new ResponseEntity<>(bgservice.addBoardGame(boardGame), HttpStatus.CREATED);//201
     }
 
     @PutMapping("boardgames")
-    public BoardGame editBoardGame(@RequestBody BoardGame boardGame){
-        return bgservice.updateBoardGame(boardGame);
+    public ResponseEntity<BoardGame> editBoardGame(@RequestBody BoardGame boardGame){
+        if(!bgservice.boardGameExistById(boardGame.getId())){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);//404 not there
+        }
+        return new ResponseEntity<>(bgservice.updateBoardGame(boardGame), HttpStatus.OK);//200
     }
 
     @DeleteMapping("boardgames")
-    public void  deleteBoardGame(@RequestBody BoardGame boardGame){
+    public ResponseEntity deleteBoardGame(@RequestBody BoardGame boardGame){
+        if(!bgservice.boardGameExistById(boardGame.getId())){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);//404 not there
+        }
         bgservice.deleteBoardGame(boardGame.getId());
+        return new ResponseEntity(HttpStatus.NO_CONTENT);//204 deleted
     }
 }
